# storage/api-docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Nutrients API",
        "description": "REST API for managing ingredients and their nutritional data.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/nutrients": {
            "get": {
                "tags": [
                    "Nutrients"
                ],
                "summary": "List nutrients (paginated)",
                "operationId": "59153af12457b12724a580787f70c96b",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of nutrients",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Nutrient"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 100
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 25
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 4
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "frontend": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Nutrients"
                ],
                "summary": "Create a new nutrient",
                "operationId": "606e7a95222294837bdd4c6e76780226",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "source_id",
                                    "name"
                                ],
                                "properties": {
                                    "source_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Vitamin C"
                                    },
                                    "external_id": {
                                        "type": "string",
                                        "example": "1004",
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "parent_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "vitamin-c",
                                        "nullable": true
                                    },
                                    "canonical_unit_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "iu_to_canonical_factor": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    },
                                    "is_label_standard": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "display_order": {
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Nutrient"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "frontend": []
                    }
                ]
            }
        },
        "/api/nutrients/{nutrient}": {
            "get": {
                "tags": [
                    "Nutrients"
                ],
                "summary": "Get a single nutrient with all relationships",
                "operationId": "bedeae6923057eede2de724c1539b771",
                "parameters": [
                    {
                        "name": "nutrient",
                        "in": "path",
                        "description": "Nutrient ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Nutrient resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Nutrient"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "frontend": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Nutrients"
                ],
                "summary": "Update a nutrient",
                "operationId": "29dd63bc25f96eb426c742025fa497a0",
                "parameters": [
                    {
                        "name": "nutrient",
                        "in": "path",
                        "description": "Nutrient ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "source_id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "external_id": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "parent_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "slug": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "canonical_unit_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "iu_to_canonical_factor": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    },
                                    "is_label_standard": {
                                        "type": "boolean"
                                    },
                                    "display_order": {
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Nutrient"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "frontend": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Nutrients"
                ],
                "summary": "Soft-delete a nutrient",
                "operationId": "85ab7fd539285e785253af7299131755",
                "parameters": [
                    {
                        "name": "nutrient",
                        "in": "path",
                        "description": "Nutrient ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "frontend": []
                    }
                ]
            }
        },
        "/api/nutrients/search": {
            "post": {
                "tags": [
                    "Nutrients"
                ],
                "summary": "Full-text search nutrients",
                "operationId": "88f86f852f92e5f4751fd9e29d7f124c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "query"
                                ],
                                "properties": {
                                    "query": {
                                        "type": "string",
                                        "example": "vitamin c"
                                    },
                                    "page": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Search results",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "hits": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Nutrient"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "page": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "frontend": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Nutrient": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "external_id": {
                        "type": "string",
                        "example": "1004",
                        "nullable": true
                    },
                    "name": {
                        "type": "string",
                        "example": "Vitamin C"
                    },
                    "description": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "slug": {
                        "type": "string",
                        "example": "vitamin-c"
                    },
                    "iu_to_canonical_factor": {
                        "type": "number",
                        "format": "float",
                        "example": null,
                        "nullable": true
                    },
                    "is_label_standard": {
                        "type": "boolean",
                        "example": true
                    },
                    "display_order": {
                        "type": "integer",
                        "example": 10,
                        "nullable": true
                    },
                    "sync_status": {
                        "type": "string",
                        "example": "synced"
                    },
                    "source": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Source"
                            }
                        ],
                        "nullable": true
                    },
                    "canonical_unit": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Unit"
                            }
                        ],
                        "nullable": true
                    },
                    "parent": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Nutrient"
                            }
                        ],
                        "nullable": true
                    },
                    "children": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Nutrient"
                        }
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/NutrientTag"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "NutrientTag": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Fat-soluble"
                    },
                    "slug": {
                        "type": "string",
                        "example": "fat-soluble"
                    },
                    "description": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Source": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "USDA FoodData Central"
                    },
                    "slug": {
                        "type": "string",
                        "example": "usda-fooddata-central"
                    },
                    "url": {
                        "type": "string",
                        "example": "https://fdc.nal.usda.gov",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Unit": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Gram"
                    },
                    "abbreviation": {
                        "type": "string",
                        "example": "g"
                    },
                    "type": {
                        "type": "string",
                        "example": "mass",
                        "nullable": true
                    },
                    "to_base_factor": {
                        "type": "number",
                        "format": "float",
                        "example": 1,
                        "nullable": true
                    },
                    "base_unit_id": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "frontend": {
                "type": "http",
                "description": "Bearer token obtained from POST /api/auth/login. Also requires X-Refresh-Token, X-Application-Name, and X-Client-Url headers on every request.",
                "scheme": "bearer"
            }
        }
    },
...
}